+++ girder/girder_large_image/web_client/views/itemViewCodemirror.js
@@ -155,6 +155,7 @@ const Formats = {
 };
 Formats['application/vnd.geo+json'] = Formats['application/json'];
 Formats['text/x-yaml'] = Formats['text/yaml'];
+Formats['application/x-yaml'] = Formats['text/yaml'];
 
 function lintGirderIni(text, callback) {
     return restRequest({