+++ girder/girder_large_image/web_client/webpack.helper.js
@@ -3,6 +3,7 @@ const path = require('path');
 const webpack = require('webpack');
 
 const CopyWebpackPlugin = require('copy-webpack-plugin');
+const {VueLoaderPlugin} = require('vue-loader');
 
 module.exports = function (config) {
     config.plugins.push(
@@ -11,7 +12,7 @@ module.exports = function (config) {
             to: path.join(config.output.path, 'extra', 'geojs.js'),
             toType: 'file'
         }, {
-            from: path.join(path.resolve(__dirname), 'node_modules', 'sinon', 'pkg', 'sinon.js'),
+            from: require.resolve('sinon/pkg/sinon.js'),
             to: path.join(config.output.path, 'extra', 'sinon.js')
         }])
     );
@@ -20,5 +21,19 @@ module.exports = function (config) {
             BUILD_TIMESTAMP: Date.now()
         })
     );
+    config.module.rules.push({
+        resource: {
+            test: /\.vue$/
+        },
+        use: [
+            require.resolve('vue-loader')
+        ]
+    });
+    config.resolve = {
+        alias: {
+            vue: process.env.NODE_ENV === 'production' ? 'vue/dist/vue.min.js' : 'vue/dist/vue.js'
+        }
+    };
+    config.plugins.push(new VueLoaderPlugin());
     return config;
 };